Create `init.rs`
